+++ changelog/66056.changed.md
@@ -1,0 +1,1 @@
+re-work the aptpkg module to remove system libraries that onedir and virtualenvs do not have access. Streamline testing, and code use to needed libraries only.
