+++ changelog/66201.added.md
@@ -1,0 +1,1 @@
+added pkg.which to aptpkg, for finding which package installed a file.
